+++ tm-server/src/main/java/tm/server/repository/mybatis/mapper/TaskMapper.java
@@ -9,16 +9,17 @@ import java.util.List;
 public interface TaskMapper {
 
     String FIND_ALL  = "SELECT * FROM `task`";
-    String FIND_ALL_SORTED  = "SELECT * FROM `task` ORDER BY #{sortMethod}";
     String FIND_BY_USER_ID  = "SELECT * FROM `task` WHERE `userId` = #{userId}";
-    String FIND_BY_PROJECT_ID  = "SELECT * FROM `task` WHERE `projectId` = #{projectId}";
+    String FIND_BY_USER_ID_SORTED  = "SELECT * FROM `task` WHERE `userId` = #{userId} ORDER BY #{sortMethod}";
+    String FIND_BY_PROJECT_ID  = "SELECT * FROM `task` WHERE `userId` = #{userId} AND `projectId` = #{projectId}";
     String FIND_BY_NAME  = "SELECT * FROM `task` WHERE `userId` = #{userId} AND `name` = #{name}";
     String FIND_BY_NAME_SORTED  = "SELECT * FROM `task` WHERE `userId` = #{userId} AND `name` = #{name} ORDER BY #{sortMethod}";
-    String FIND_BY_ID  = "SELECT * FROM `task` WHERE `id` = #{id}";
+    String FIND_BY_ID  = "SELECT * FROM `task` WHERE `userId` = #{userId} AND `id` = #{id}";
+    String SEARCH = "SELECT * FROM `task` WHERE ( `userId` = #{userId} ) AND ( `name` LIKE CONCAT(\"%\", #{line}, \"%\") OR `description` LIKE CONCAT(\"%\", #{line}, \"%\") )";
     String PERSIST  = "INSERT INTO `task` (`id`, `userId`, `projectId`, `status`, `name`, `description`, `creationDate`, `startDate`, `endDate`) " +
             "VALUES (#{id}, #{userId}, #{projectId}, #{status}, #{name}, #{description}, #{creationDate}, #{startDate}, #{endDate})";
-    String REMOVE_BY_ID  = "DELETE FROM `task` WHERE `id` = #{id}";
-    String REMOVE_BY_PROJECT_ID  = "DELETE FROM `task` WHERE `projectId` = #{projectId}";
+    String REMOVE_BY_ID  = "DELETE FROM `task` WHERE `userId` = #{userId} AND `id` = #{id}";
+    String REMOVE_BY_PROJECT_ID  = "DELETE FROM `task` WHERE `userId` = #{userId} AND `projectId` = #{projectId}";
     String REMOVE_BY_NAME  = "DELETE FROM `task`, WHERE `userId` = #{userId} AND `name` = #{name}";
     String REMOVE_ALL  = "DELETE FROM `task`, WHERE `userId` = #{userId}";
 
@@ -35,14 +36,14 @@ public interface TaskMapper {
     @Select(FIND_ALL)
     List<Task> findAll() throws Exception;
 
-    @Select(FIND_ALL_SORTED)
+    @Select(FIND_BY_USER_ID)
     List<Task> findByUserId(@NotNull String userId) throws Exception;
 
-    @Select(FIND_BY_USER_ID)
-    List<Task> findAllAndSort(@NotNull String sortMethod) throws Exception;
+    @Select(FIND_BY_USER_ID_SORTED)
+    List<Task> findByUserIdAndSort(@NotNull @Param("userId") String userId, @NotNull @Param("sortMethod") String sortMethod) throws Exception;
 
     @Select(FIND_BY_PROJECT_ID)
-    List<Task> findByProjectId(@NotNull String projectId) throws Exception;
+    List<Task> findByProjectId(@NotNull @Param("userId") String userId, @NotNull @Param("projectId") String projectId) throws Exception;
 
     @Select(FIND_BY_NAME)
     List<Task> findByName(@NotNull @Param("userId") String userId, @NotNull @Param("name") String name) throws Exception;
@@ -51,16 +52,19 @@ public interface TaskMapper {
     List<Task> findByNameAndSort(@NotNull @Param("userId") String userId, @NotNull @Param("name") String name, @NotNull @Param("sortMethod") String sortMethod) throws Exception;
 
     @Select(FIND_BY_ID)
-    Task findOne(@NotNull String id) throws Exception;
+    Task findOne(@NotNull @Param("userId") String userId, @NotNull @Param("id") String id) throws Exception;
+
+    @Select(SEARCH)
+    List<Task> search(@NotNull @Param("userId") String userId, @NotNull @Param("line") String line);
 
     @Insert(PERSIST)
     void persist(@NotNull Task project) throws Exception;
 
     @Delete(REMOVE_BY_ID)
-    void removeById(@NotNull String id) throws Exception;
+    void removeById(@NotNull @Param("userId") String userId, @NotNull @Param("id") String id) throws Exception;
 
     @Delete(REMOVE_BY_PROJECT_ID)
-    void removeByProjectId(@NotNull String projectId) throws Exception;
+    void removeByProjectId(@NotNull @Param("userId") String userId, @NotNull @Param("projectId") String projectId) throws Exception;
 
     @Delete(REMOVE_BY_NAME)
     void removeByName(@NotNull @Param("userId") String userId, @NotNull @Param("name") String name) throws Exception;
